put it back

+++ docker/run-mysql.sh
@@ -3,17 +3,26 @@
 # The CloudSQL console simply states "MySQL 5.6" so we may not match the minor version number
 MYSQL_VERSION=5.6
 start() {
-    echo "starting up mysql container..."
-    docker run --name $CONTAINER -e MYSQL_ROOT_PASSWORD=leonardo-test -e MYSQL_USER=leonardo-test -e MYSQL_PASSWORD=leonardo-test -e MYSQL_DATABASE=testdb -d -p 3310:3306 mysql/mysql-server:$MYSQL_VERSION
 
-    # validate mysql
-    echo "running mysql validation..."
-    docker run --rm --link $CONTAINER:mysql -v $PWD/docker/sql_validate.sh:/working/sql_validate.sh broadinstitute/dsde-toolbox /working/sql_validate.sh leonardo
-    if [ 0 -eq $? ]; then
-        echo "mysql validation succeeded."
-    else
-        echo "mysql validation failed."
-        exit 1
+    RUNNING=$(docker inspect -f {{.State.Running}} $CONTAINER || echo "false")
+
+    if ! $RUNNING; then
+        echo "attempting to remove old $CONTAINER container..."
+        docker rm -f $CONTAINER || echo "docker rm failed. nothing to rm."
+
+        # start up mysql
+        echo "starting up mysql container..."
+        docker run --name $CONTAINER -e MYSQL_ROOT_PASSWORD=leonardo-test -e MYSQL_USER=leonardo-test -e MYSQL_PASSWORD=leonardo-test -e MYSQL_DATABASE=testdb -d -p 3310:3306 mysql/mysql-server:$MYSQL_VERSION
+
+        # validate mysql
+        echo "running mysql validation..."
+        docker run --rm --link $CONTAINER:mysql -v $PWD/docker/sql_validate.sh:/working/sql_validate.sh broadinstitute/dsde-toolbox /working/sql_validate.sh leonardo
+        if [ 0 -eq $? ]; then
+            echo "mysql validation succeeded."
+        else
+            echo "mysql validation failed."
+            exit 1
+        fi
     fi
 }
 
@@ -23,7 +32,7 @@ stop() {
     docker rm -v $CONTAINER || echo "mysql rm -v failed.  container already destroyed."
 }
 
-CONTAINER=mysql-$(date +%s)
+CONTAINER=mysql
 COMMAND=$1
 
 if [ ${#@} == 0 ]; then
